requirements.txt
pandas==2.1.3
pymongo==4.6.0
cloudinary==1.35.0
python-dotenv==1.0.0
tqdm==4.66.1
openpyxl==3.10.10
